removed unnecessary casts

+++ main.cpp
@@ -252,14 +252,14 @@ void part3()
     IntType it ( 34 );
     DoubleType pi( 3.14 );
 
-    std::cout << "The result of FloatType^4 divided by IntType is: " << static_cast<float>(ft.multiply( static_cast<float>(ft) ).multiply( static_cast<float>(ft) ).divide( static_cast<int>(it) ) ) << std::endl;
-    std::cout << "The result of DoubleType times 3 plus IntType is : " << static_cast<double>(dt.multiply( static_cast<double>(3) ).add( static_cast<int>(it) ) ) << std::endl;
-    std::cout << "The result of IntType divided by 3.14 multiplied by DoubleType minus FloatType is: " << static_cast<int>(it.divide( static_cast<int>(pi) ).multiply( static_cast<int>(dt) ).subtract( static_cast<int>(ft) ) ) << std::endl;
+    std::cout << "The result of FloatType^4 divided by IntType is: " << ft.multiply( static_cast<float>(ft) ).multiply( static_cast<float>(ft) ).divide( static_cast<int>(it) ) << std::endl;
+    std::cout << "The result of DoubleType times 3 plus IntType is : " << dt.multiply( static_cast<double>(3) ).add( static_cast<int>(it) ) << std::endl;
+    std::cout << "The result of IntType divided by 3.14 multiplied by DoubleType minus FloatType is: " << it.divide( static_cast<int>(pi) ).multiply( static_cast<int>(dt) ).subtract( static_cast<int>(ft) ) << std::endl;
     std::cout << "An operation followed by attempts to divide by 0, which are ignored and warns user: " << std::endl;
-    std::cout << static_cast<int>(it.multiply( static_cast<int>(it) ).divide(0).divide( static_cast<int>(0.0f) ).divide( static_cast<int>(0.0) ) )<<std::endl;
+    std::cout << it.multiply( static_cast<int>(it) ).divide(0).divide( static_cast<int>(0.0f) ).divide( static_cast<int>(0.0) ) <<std::endl;
     
-    std::cout << "FloatType x IntType  =  " << static_cast<int>(it.multiply( static_cast<int>(ft) ) )<< std::endl;
-    std::cout << "(IntType + DoubleType + FloatType) x 24 = " << static_cast<int>(it.add( static_cast<int>(dt) ).add( static_cast<int>(ft) ).multiply( 24 ) ) << std::endl;
+    std::cout << "FloatType x IntType  =  " << it.multiply( static_cast<int>(ft) )<< std::endl;
+    std::cout << "(IntType + DoubleType + FloatType) x 24 = " << it.add( static_cast<int>(dt) ).add( static_cast<int>(ft) ).multiply( 24 ) << std::endl;
 }
 
 int main()
@@ -269,44 +269,44 @@ int main()
     DoubleType dt ( 2 );
     IntType it ( 2 ) ;
 
-    std::cout << "FloatType add result=" << ( static_cast<float>( ft.add( 2.0f )) ) << std::endl;
-    std::cout << "FloatType subtract result=" << ( static_cast<float>( ft.subtract( 2.0f )) ) << std::endl;
-    std::cout << "FloatType multiply result=" << ( static_cast<float>( ft.multiply( 2.0f )) ) << std::endl;
-    std::cout << "FloatType divide result=" << ( static_cast<float>(ft.divide( 16.0f)) ) << std::endl << std::endl;
+    std::cout << "FloatType add result=" << ( ft.add( 2.0f ) ) << std::endl;
+    std::cout << "FloatType subtract result=" << ( ft.subtract( 2.0f ) ) << std::endl;
+    std::cout << "FloatType multiply result=" << ( ft.multiply( 2.0f ) ) << std::endl;
+    std::cout << "FloatType divide result=" << ( ft.divide( 16.0f) ) << std::endl << std::endl;
 
-    std::cout << "DoubleType add result=" << ( static_cast<double>(dt.add(2.0)) ) << std::endl;
-    std::cout << "DoubleType subtract result=" << ( static_cast<double>(dt.subtract(2.0)) ) << std::endl;
-    std::cout << "DoubleType multiply result=" << ( static_cast<double>(dt.multiply(2.0)) ) << std::endl;
-    std::cout << "DoubleType divide result=" << ( static_cast<double>(dt.divide( static_cast<double>(5.f)) ) ) << std::endl << std::endl;
+    std::cout << "DoubleType add result=" << ( dt.add(2.0) ) << std::endl;
+    std::cout << "DoubleType subtract result=" << ( dt.subtract(2.0) ) << std::endl;
+    std::cout << "DoubleType multiply result=" << ( dt.multiply(2.0) ) << std::endl;
+    std::cout << "DoubleType divide result=" << ( dt.divide( static_cast<double>(5.f)) ) << std::endl << std::endl;
 
-    std::cout << "IntType add result=" << ( static_cast<int>(it.add(2)) ) << std::endl;
-    std::cout << "IntType subtract result=" << ( static_cast<int>(it.subtract(2)) ) << std::endl;
-    std::cout << "IntType multiply result=" << ( static_cast<int>(it.multiply(2)) ) << std::endl;
-    std::cout << "IntType divide result=" << ( static_cast<int>(it.divide(3)) ) << std::endl << std::endl;
-    std::cout << "Chain calculation = " << ( static_cast<int>(it.multiply(1000).divide(2).subtract(10).add(100)) ) << std::endl;
+    std::cout << "IntType add result=" << ( it.add(2) ) << std::endl;
+    std::cout << "IntType subtract result=" << ( it.subtract(2) ) << std::endl;
+    std::cout << "IntType multiply result=" << ( it.multiply(2) ) << std::endl;
+    std::cout << "IntType divide result=" << ( it.divide(3) ) << std::endl << std::endl;
+    std::cout << "Chain calculation = " << ( it.multiply(1000).divide(2).subtract(10).add(100) ) << std::endl;
 
         // FloatType object instanciation and method tests
     // --------
-    std::cout << "New value of ft = (ft + 3.0f) * 1.5f / 5.0f = " << ( static_cast<float>(ft.add(3.0f).multiply(1.5f).divide(5.0f)) ) << std::endl;
+    std::cout << "New value of ft = (ft + 3.0f) * 1.5f / 5.0f = " << ( ft.add(3.0f).multiply(1.5f).divide(5.0f))  << std::endl;
        
     std::cout << "---------------------\n" << std::endl; 
     
     // DoubleType/IntType object instanciation and method tests
     // --------
-    std::cout << "Initial value of dt: " << static_cast<double>(dt) << std::endl;
-    std::cout << "Initial value of it: " << static_cast<int>(it) << std::endl;
+    std::cout << "Initial value of dt: " << dt << std::endl;
+    std::cout << "Initial value of it: " << it << std::endl;
     // --------
     std::cout << "Use of function concatenation (mixed type arguments) " << std::endl;
-    std::cout << "New value of dt = (dt * it) / 5.0f + ft = " << ( static_cast<double>( dt.multiply( static_cast<int>(it) ).divide( static_cast<double>(5.0f) ).add( static_cast<double>(ft)) ) ) << std::endl;
+    std::cout << "New value of dt = (dt * it) / 5.0f + ft = " << ( dt.multiply( static_cast<int>(it) ).divide( static_cast<double>(5.0f) ).add( static_cast<double>(ft)) ) << std::endl;
 
     std::cout << "---------------------\n" << std::endl; 
     
     // Intercept division by 0
     // --------
     std::cout << "Intercept division by 0 " << std::endl;
-    std::cout << "New value of it = it / 0 = " << ( static_cast<int>(it.divide(0)) ) << std::endl;
-    std::cout << "New value of ft = ft / 0 = " << ( static_cast<float>(ft.divide(0)) ) << std::endl;
-    std::cout << "New value of dt = dt / 0 = " << ( static_cast<double>(dt.divide(0)) ) << std::endl;
+    std::cout << "New value of it = it / 0 = " << ( it.divide(0) ) << std::endl;
+    std::cout << "New value of ft = ft / 0 = " << ( ft.divide(0) ) << std::endl;
+    std::cout << "New value of dt = dt / 0 = " << ( dt.divide(0) ) << std::endl;
 
     std::cout << "---------------------\n" << std::endl;
